+++ Market.Models/ViewModels/UrunViewModel.cs
@@ -11,7 +11,7 @@ namespace Market.Models.ViewModels
     {
         public int UrunId { get; set; }
         public string UrunAd { get; set; }
-        public short UrunStok { get; set; }
+        public int UrunStok { get; set; }
         public decimal UrunFiyat { get; set; }
         public int KategoriId { get; set; }
